Fix am/rebase edge-case
I'm unclear on the circumstances that cause this situation but a rebase
action, fairly routine to my mind, incorrectly registered as an am
action (and then murdered the action's state). This change nearly copies
a section from __git_ps1; it assumes that the ambiguous "AM/REBASE"
state should resolve rebase.